requirements.txt
cocos2d==0.6.3
pyglet==1.2.4
six==1.10.0
